add: creating playerHealthScript
