+++ Assets/Prefabs/HeartsContainerPrefab.prefab
@@ -75,5 +75,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e0b96239ee5dcf084be36a8c71a7f8a7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  status: {fileID: 0}
+  healthStatus: {fileID: 0}
   heartPrefab: {fileID: 3803928884588761936, guid: f30bcf54915cf819c99c54c864626247, type: 3}
