+++ Assets/Scenes/GameScene.unity
@@ -276,7 +276,10 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 6219646412613573617, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+      insertIndex: 5
+      addedObject: {fileID: 2080808166}
   m_SourcePrefab: {fileID: 100100000, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
 --- !u!95 &346729848 stripped
 Animator:
@@ -393,6 +396,11 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &1431991378 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 6219646412613573617, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
+  m_PrefabInstance: {fileID: 346729847}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1731713988
 GameObject:
   m_ObjectHideFlags: 0
@@ -596,12 +604,26 @@ MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 571736130042948404, guid: cd36a56cee42091f2888f50a3ff46fa3, type: 3}
   m_PrefabInstance: {fileID: 346729847}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
+  m_GameObject: {fileID: 1431991378}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: ef0863be3f918582b90f577d24107617, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &2080808166
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1431991378}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 75e4775abe4417dba85ad5403bc2ebc7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHealth: 6
+  currentHealth: 0
 --- !u!1 &2138874834
 GameObject:
   m_ObjectHideFlags: 0
