+++ Assets/Scripts/Player/Status.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 921c4ebd9d0492b02979470519689b17
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
