+++ Assets/Scripts/UI/Hearts/HeartUIManager.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 // se refere ao containner de hears
 public class HeartUIManager : MonoBehaviour
 {
-    [SerializeField] private PlayerStatus status; //pq na unity eu tive que colocar o prefab de player e não o arquivo de script?
+    [SerializeField] private PlayerHealthStatus healthStatus; //pq na unity eu tive que colocar o prefab de player e não o arquivo de script?
     [SerializeField] private GameObject heartPrefab;
     List<HealtHeart> hearts = new List<HealtHeart>(); //pq e uma lista de scripts e nao de game object?
 
@@ -22,12 +22,12 @@ public class HeartUIManager : MonoBehaviour
     }
     private void OnEnable()
     {
-        PlayerStatus.OnPlayerDamaged += DrawHearts;
+        PlayerHealthStatus.OnPlayerDamaged += DrawHearts;
     }
 
     private void OnDesable()
     {
-        PlayerStatus.OnPlayerDamaged -= DrawHearts;
+        PlayerHealthStatus.OnPlayerDamaged -= DrawHearts;
     }
 
 
@@ -40,7 +40,7 @@ public class HeartUIManager : MonoBehaviour
 
     private void CreateInitialEmptyHearts()
     {
-        int maxHealth = status.MaxHealth;
+        int maxHealth = healthStatus.MaxHealth;
         float maxHealthRemainder = maxHealth % 2;
         int heartsToMake = (int)(maxHealth / 2 + maxHealthRemainder);
         for (int i = 0; i < heartsToMake; i++)
@@ -49,7 +49,7 @@ public class HeartUIManager : MonoBehaviour
 
     private void CreateHealthHearts()
     {
-        int currentHealth = status.CurrentHealth;
+        int currentHealth = healthStatus.CurrentHealth;
         for (int i = 0; i < hearts.Count; i++)
         {
             int heartStatusRemainder = Mathf.Clamp(currentHealth - (i * 2), 0, 2);
